creates component for View Team
